+++ src/main/java/net/chikaboom/controller/rest/ServiceRestController.java
@@ -1,16 +1,13 @@
 package net.chikaboom.controller.rest;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.github.fge.jsonpatch.JsonPatch;
-import com.github.fge.jsonpatch.JsonPatchException;
 import lombok.RequiredArgsConstructor;
-import net.chikaboom.exception.NoSuchDataException;
+import net.chikaboom.model.database.Account;
 import net.chikaboom.model.database.Service;
 import net.chikaboom.service.data.ServiceDataService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -18,31 +15,51 @@ import java.util.Optional;
 
 @RequiredArgsConstructor
 @RestController
-@RequestMapping()
 public class ServiceRestController {
 
-    private final ObjectMapper objectMapper;
     private final ServiceDataService serviceDataService;
 
-    @GetMapping("/{idService}")
+    @PreAuthorize("permitAll()")
+    @GetMapping("/services/{idService}")
     public ResponseEntity<Service> findService(@PathVariable int idService) {
         Optional<Service> serviceOptional = serviceDataService.findById(idService);
 
         return serviceOptional.map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
-
     }
 
-    @GetMapping
+    @PreAuthorize("permitAll()")
+    @GetMapping("/services")
     public ResponseEntity<List<Service>> findAllServices() {
-        return new ResponseEntity<>(serviceDataService.findAll(), HttpStatus.OK);
+        return ResponseEntity.ok(serviceDataService.findAll());
     }
 
-    @PostMapping
+    /**
+     * Создает данные об услуге мастера
+     *
+     * @param service услуга, которую предоставляет мастер
+     * @return обновленная услуга
+     */
+    @PreAuthorize("hasRole('MASTER')")
+    @PostMapping("/services")
     public ResponseEntity<Service> createService(@RequestBody Service service) {
+        Account masterAccount = service.getAccount();
+        Account authorizedAccount = (Account) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if (authorizedAccount.getIdAccount() != masterAccount.getIdAccount()) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+
         return ResponseEntity.ok(serviceDataService.create(service));
     }
 
-    @PutMapping("/{idService}")
+    /**
+     * Обновляет данные об услуге мастера
+     *
+     * @param service услуга, которую предоставляет мастер
+     * @return обновленная услуга
+     */
+    @PreAuthorize("hasRole('MASTER')")
+    @PutMapping("/services/{idService}")
     public ResponseEntity<Service> replaceService(@PathVariable int idService, @RequestBody Service service) {
         Optional<Service> serviceOptional = serviceDataService.findById(idService);
 
@@ -50,32 +67,54 @@ public class ServiceRestController {
             return ResponseEntity.notFound().build();
         }
 
+        Account masterAccount = service.getAccount();
+        Account authorizedAccount = (Account) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if (authorizedAccount.getIdAccount() != masterAccount.getIdAccount()) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+
+        service.setIdService(idService);
+
         return ResponseEntity.ok(serviceDataService.update(service));
     }
 
-    @PatchMapping(path = "/{idService}", consumes = "application/json-patch+json")
-    public ResponseEntity<Service> changeService(@PathVariable int idService, @RequestBody JsonPatch jsonPatch) {
-        try {
-            Service service = serviceDataService.findById(idService)
-                    .orElseThrow(() -> new NoSuchDataException("Service with id " + idService + " not found."));
-            Service patchedService = applyPatchToService(jsonPatch, service);
+    /**
+     * Удаляет из базы данных указанную услугу
+     *
+     * @param idService идентификатор услуги, которую необходимо удалить
+     * @return строку, содержащую результат удаления
+     */
+    @PreAuthorize("hasRole('MASTER')")
+    @DeleteMapping("/services/{idService}")
+    public ResponseEntity<Object> deleteService(@PathVariable int idService) {
+        Optional<Service> serviceOptional = serviceDataService.findById(idService);
 
-            return ResponseEntity.ok(serviceDataService.update(patchedService));
-        } catch (JsonPatchException | JsonProcessingException e) {
-            return ResponseEntity.internalServerError().build();
-        } catch (NoSuchDataException e) {
+        if (!serviceOptional.isPresent()) {
             return ResponseEntity.notFound().build();
         }
-    }
 
-    @DeleteMapping("/{idService}")
-    public void deleteService(@PathVariable int idService) {
+        Account masterAccount = serviceOptional.get().getAccount();
+        Account authorizedAccount = (Account) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if (authorizedAccount.getIdAccount() != masterAccount.getIdAccount()) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+
         serviceDataService.deleteById(idService);
+
+        return ResponseEntity.ok().build();
     }
 
-    private Service applyPatchToService(JsonPatch jsonPatch, Service targetService)
-            throws JsonProcessingException, JsonPatchException {
-        JsonNode patched = jsonPatch.apply(objectMapper.convertValue(targetService, JsonNode.class));
-        return objectMapper.treeToValue(patched, Service.class);
+    /**
+     * Загружает информацию обо всех услугах, который создал пользователь (idAccount - идентифицирует пользователя).
+     *
+     * @param idAccount идентификатор аккаунта
+     * @return полную информацию в формате JSON обо всех созданных пользователем услугах.
+     */
+    @PreAuthorize("permitAll()")
+    @GetMapping("/accounts/{idAccount}/services")
+    public List<Service> findAllServicesById(@PathVariable int idAccount) {
+        return serviceDataService.findAllServicesByIdAccount(idAccount);
     }
 }
